+++ src/lib/forms-schema/index.ts
@@ -231,6 +231,34 @@ export const paymentEventTypes: SubmissionEventTypes.FormPaymentEventType[] = [
   'WESTPAC_QUICK_STREAM',
   'NSW_GOV_PAY',
 ]
+
+const formPaymentEventAmountConfiguration = {
+  amountType: Joi.string().valid('FORM_ELEMENT', 'NUMBER', 'EXPRESSION'),
+  elementId: Joi.when('amountType', {
+    switch: [
+      {
+        is: 'NUMBER',
+        then: Joi.any().strip(),
+      },
+      {
+        is: 'EXPRESSION',
+        then: Joi.any().strip(),
+      },
+    ],
+    otherwise: Joi.string().required(),
+  }),
+  paymentAmount: Joi.when('amountType', {
+    is: 'NUMBER',
+    then: Joi.number().required(),
+    otherwise: Joi.any().strip(),
+  }),
+  paymentCalculation: Joi.when('amountType', {
+    is: 'EXPRESSION',
+    then: Joi.string().required(),
+    otherwise: Joi.any().strip(),
+  }),
+}
+
 export const PaymentEventSchema = Joi.object({
   type: Joi.string()
     .required()
@@ -240,7 +268,7 @@ export const PaymentEventSchema = Joi.object({
     .when('type', {
       is: 'BPOINT',
       then: Joi.object().keys({
-        elementId: Joi.string().required(),
+        ...formPaymentEventAmountConfiguration,
         environmentId: Joi.string().uuid().required(),
         crn2: Joi.string(),
         crn3: Joi.string(),
@@ -249,7 +277,7 @@ export const PaymentEventSchema = Joi.object({
     .when('type', {
       is: Joi.valid('WESTPAC_QUICK_STREAM'),
       then: Joi.object().keys({
-        elementId: Joi.string().required(),
+        ...formPaymentEventAmountConfiguration,
         environmentId: Joi.string().uuid().required(),
         customerReferenceNumber: Joi.string().required(),
       }),
@@ -257,14 +285,14 @@ export const PaymentEventSchema = Joi.object({
     .when('type', {
       is: 'CP_PAY',
       then: Joi.object().keys({
-        elementId: Joi.string().required(),
+        ...formPaymentEventAmountConfiguration,
         gatewayId: Joi.string().uuid().required(),
       }),
     })
     .when('type', {
       is: 'NSW_GOV_PAY',
       then: Joi.object().keys({
-        elementId: Joi.string().required(),
+        ...formPaymentEventAmountConfiguration,
         primaryAgencyId: Joi.string().uuid().required(),
         productDescription: Joi.string().required().max(250),
         customerReference: Joi.string().max(250),